extract_backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
rapidfuzz
google-generativeai
